Add bitgenerator for the ChaCha family of PRNG's.
This PR adds a bitgenerator for the ChaCha family of PRNG's.
It defaults to the lowest number of rounds that pass stringent
randomness statistical tests (i.e the TestU01 suite). This member
of the chacha family is `ChaCha<4>` as highlighed by two blog posts
[here](https://www.johndcook.com/blog/2019/03/03/do-the-chacha/) and
[there](https://www.pcg-random.org/posts/more-practrand-passes.html).
A `initialize_full` function is provided to allow a user to specify a
custom number of rounds, the key array as well as the initial counter value.